requirements.txt
aiofile==3.5.1
aiohttp==3.7.4.post0
altgraph==0.17
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
cachetools==4.2.2
caio==0.8.4
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
colorlog==5.0.1
cssselect==1.1.0
discord.py==1.7.2
docstring-parser==0.8.1
fake-useragent==0.1.11
future==0.18.2
httplib2==0.19.1
idna==2.10
isodate==0.6.0
lxml==4.6.3
multidict==5.1.0
mysql-connector-python==8.0.26
parse==1.19.0
pefile==2021.5.24
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyee==8.1.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pymongo==3.12.0
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
requests==2.25.1
requests-html==0.10.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
tqdm==4.61.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
w3lib==1.22.0
websockets==8.1
yarl==1.6.3
